+++ src/__tests__/MonoStore.ts
@@ -16,15 +16,15 @@ export default class MonoStore {
 
   @task.async
   async addAsync(num: number) {
+    console.log("3")
     await delay(100);
-    console.log("addAsync")
+    console.log("4")
     this.cnt += num;
   }
 
   @task.async
   async task() {
     this.add(10);
-    console.log("add 10")
     await this.addAsync(5);
     this.cnt *= 2;
   }
@@ -32,7 +32,6 @@ export default class MonoStore {
   @task.async
   async cancelable() {
     await delay(0);
-    console.log("cancelable")
     this.cnt += 1;
   }
 }
